python/requirements.txt
# Python dependencies for invoicing scripts
openai>=1.19.0
pdf2image>=1.16.3
Pillow>=10.1.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.12
requests>=2.31.0
pdfplumber>=0.10.0
google-generativeai>=0.3.0
google-cloud-vision>=3.4.0
google-cloud-storage>=2.14.0
PyMuPDF>=1.23.0
pymongo>=4.6.0

